compiling and linking fwd/bwd passes and restructuring unit tests as we go
