+++ parmoo/tests/unit_tests/test_moop_grads.py
@@ -1,110 +1,4 @@
 
-def test_MOOP_evaluatePenalty():
-    """ Check that the MOOP class handles evaluating penalty function properly.
-
-    Initialize a MOOP object and check that the evaluatePenalty() function
-    works correctly.
-
-    """
-
-    from parmoo import MOOP
-    from parmoo.surrogates import GaussRBF
-    from parmoo.searches import LatinHypercube
-    from parmoo.optimizers import GlobalSurrogate_PS
-    import numpy as np
-
-    # Create 2 SimGroups for later
-    g1 = {'n': 3,
-          'm': 1,
-          'hyperparams': {},
-          'search': LatinHypercube,
-          'sim_func': lambda x: [np.linalg.norm(x)],
-          'surrogate': GaussRBF}
-    g2 = {'n': 3,
-          'm': 2,
-          'hyperparams': {},
-          'search': LatinHypercube,
-          'sim_func': lambda x: [np.linalg.norm(x-1.0), np.linalg.norm(x-0.5)],
-          'surrogate': GaussRBF}
-
-    # Create several differentiable functions and constraints.
-    def f1(x, s, der=0):
-        if der == 0:
-            return np.dot(x, x)
-        if der == 1:
-            return 2.0 * x
-        if der == 2:
-            return np.zeros(s.size)
-
-    def f2(x, s, der=0):
-        if der == 0:
-            return np.dot(s - 0.5, s - 0.5)
-        if der == 1:
-            return np.zeros(x.size)
-        if der == 2:
-            return 2.0 * s - 1.0
-
-    def c1(x, s, der=0):
-        if der == 0:
-            return x[0] - 0.25
-        if der == 1:
-            return np.ones(x.size)
-        if der == 2:
-            return np.zeros(x.size)
-
-    def c2(x, s, der=0):
-        if der == 0:
-            return s[0] - 0.25
-        if der == 1:
-            return np.zeros(s.size)
-        if der == 2:
-            return np.ones(s.size)
-
-    # Initialize a MOOP with 2 SimGroups and 3 objectives
-    moop1 = MOOP(GlobalSurrogate_PS)
-    for i in range(3):
-        moop1.addDesign({'lb': 0.0, 'ub': 1.0})
-    moop1.addObjective({'obj_func': f1})
-    assert (np.all(moop1.evaluatePenalty(np.zeros(3)) == np.zeros(1)))
-    assert (np.all(moop1.evaluatePenalty(np.ones(3)) == 3.0 * np.ones(1)))
-    moop1.addConstraint({'constraint': c1})
-    assert (np.all(moop1.evaluatePenalty(np.zeros(3)) == np.zeros(1)))
-    assert (np.all(moop1.evaluatePenalty(np.ones(3)) == 3.75 * np.ones(1)))
-    moop1 = MOOP(GlobalSurrogate_PS)
-    for i in range(3):
-        moop1.addDesign({'lb': 0.0, 'ub': 1.0})
-    moop1.addSimulation(g1, g2)
-    moop1.evaluateSimulation(np.ones(3), 0)
-    moop1.evaluateSimulation(np.ones(3), 1)
-    moop1.fitSurrogates()
-    moop1.addObjective({'obj_func': f1})
-    moop1.addObjective({'obj_func': f1})
-    moop1.setSurrogateTR(np.zeros(3), np.ones(3) * np.infty)
-    assert (np.all(moop1.evaluatePenalty(np.zeros(3)) == np.zeros(1)))
-    assert (np.all(moop1.evaluatePenalty(np.ones(3)) == 3.0 * np.ones(1)))
-    moop1.addConstraint({'constraint': c1})
-    moop1.setSurrogateTR(np.zeros(3), np.ones(3) * np.infty)
-    assert (np.all(moop1.evaluatePenalty(np.zeros(3)) == np.zeros(1)))
-    assert (np.all(moop1.evaluatePenalty(np.ones(3)) == 3.75 * np.ones(1)))
-    # Adjust the scaling and compare
-    moop2 = MOOP(GlobalSurrogate_PS)
-    moop2.addDesign({'lb': -1.0, 'ub': 1.0},
-                    {'lb': 0.0, 'ub': 2.0},
-                    {'lb': -0.5, 'ub': 1.5})
-    moop2.addSimulation(g1, g2)
-    moop2.evaluateSimulation(np.ones(3), 0)
-    moop2.evaluateSimulation(np.ones(3), 1)
-    moop2.fitSurrogates()
-    moop2.setSurrogateTR(np.zeros(3), np.ones(3) * np.infty)
-    moop2.addObjective({'obj_func': f1})
-    moop2.addObjective({'obj_func': f1})
-    moop2.addConstraint({'constraint': c1})
-    x = moop1.__embed__(np.ones(3))
-    xx = moop2.__embed__(np.ones(3))
-    assert (np.linalg.norm(moop1.evaluatePenalty(x) -
-                           moop2.evaluatePenalty(xx)) < 0.00000001)
-
-
 def test_MOOP_evaluateGradients_1():
     """ Check that the MOOP class handles evaluating gradients properly.
